Crate gfatk

gfatk is a tool for Graphical Fragment Assembly (GFA) manipulation.

GFA’s are at their heart, simple, directed graphs. As such, all internal representations of GFA’s are petgraph::Graph’s.

gfatk is designed mainly for reasonably small GFA files, and was designed for de-tangling and linearising plant mitochondrial genomes on the command line.

Almost all of the core functionality of gfatk resides in the gfatk::gfa module. The other modules are entry points for the command line application.

Modules

dot

Make a DOT language representation of a GFA.

extract

Extract a subgraph from a GFA.

extract_chloro

Extract the putative plastid subgraph in a GFA.

extract_mito

Extract the putative mitochondrial subgraph in a GFA.

fasta

Print all the sequences in a GFA to fasta format.

gfa

A module with all the methods to manipulate GFA’s in.

linear

Coerce a GFA into a fasta, finding the longest path through the graph.

load

Helper functions to load a GFA from a file, or read from STDIN. Modified from https://github.com/chfi/rs-gfa-utils/blob/2065b001d107ee9f5d7abe04d65ab82193fc5904/src/commands.rs

overlap

Generate overlapping sequences between segments in a GFA.

path

Extract a fasta given a path.

rename

Rename a GFA’s segment ID’s to a 1-indexed usize.

stats

Generate statistics about the input GFA file.

trim

Utility to trim a GFA of isolated nodes.

utils

Utility functions used throughout.