Struct gfatk::gfa::gfa::GFAtk

pub struct GFAtk(pub GFA<usize, OptionalFields>);

A wrapper around GFA from the gfa crate

Tuple Fields
0: GFA<usize, OptionalFields>
Implementations

impl GFAtk

pub fn into_ungraph(&self) -> Result<(GFAGraphLookups, GFAungraph)>

Returns a tuple of GFAGraphLookups (a struct of indices/node names) and an undirected GFA graph structure.

pub fn into_digraph(&self) -> Result<(GFAGraphLookups, GFAdigraph)>

Returns a tuple of GFAGraphLookups (a struct of indices/node names) and an directed GFA graph structure.

Most functionality of this binary is on directed graph structures

pub fn print_extract(&self, sequences_to_keep: Vec<usize>)

A method to print a GFA to STDOUT, given a vector of sequence ID’s to keep.

pub fn make_overlaps(&self, extend_length: usize) -> Result<Overlaps>

Returns the overlaps between all the segments in a GFA.

pub fn determine_path_overlaps(
&self,
chosen_path: &Vec<NodeIndex>,
graph_indices: GFAGraphLookups,
chosen_path_ids: &Vec<usize>
) -> Result<Vec<(usize, Orientation, usize, &str)>>

Computes the overlaps between segments as a vector of: (usize, Orientation, usize, &str) Which is a tuple of from/to, the orientation of the overlap, the extent of the overlap, and whether the overlap is at the start or end of a segment.

pub fn print_path_to_fasta(
&self,
merged_sorted_chosen_path_overlaps: IndexMap<usize, Vec<(Orientation, usize, &str)>>,
fasta_header: &str,
segments_not_in_path: Vec<usize>
) -> Result<()>

A method to print to STDOUT a fasta, given a path through the GFA.

The first segment is added first, then all subsequent segments (-overlap with previous segment).

pub fn print_sequences(&self) -> Result<()>

The internal function called when gfatk fasta is called.

Prints all segments of the GFA as-is.

fn parse_coverage_opt(opt: &OptFieldVal) -> Result<&f32>

Two internal functions below to parse coverage of a GFA segment.

Used in gfatk stats.

fn get_coverage(&self) -> Result<f32>

pub fn node_seq_len_and_cov(&self, node: usize) -> Result<(usize, f32)>

Return the coverage and sequence length for a segment, given a segment name.

Note segment names are always usize.

pub fn sequence_stats(&self, further: bool) -> Result<(f32, f32, usize)>

The internal function called in gfatk stats.

Returns average GC%, average coverage, and total sequence length for a GFA (sub)graph.

pub fn gen_cov_hash(
&self,
graph_lookup: &GFAGraphLookups
) -> Result<HashMap<NodeIndex, usize>>

Returns a HashMap of relative coverage of each node (segment) in the GFA.

Relative here indicates that each segment coverage is divided by the lowest coverage node, and rounded.