Function gfatk::gfa::graph::all_paths

pub fn all_paths<T, U, Ix: IndexType>(
    graph: &Graph<T, U, Directed, Ix>,
    start_node: NodeIndex<Ix>,
    end_node: NodeIndex<Ix>,
    rel_coverage_map: Option<&HashMap<NodeIndex<Ix>, usize>>,
    index: usize
) -> Result<Vec<Vec<NodeIndex<Ix>>>>

A function generic over certain types of Directed petgraph Graphs.

Given a graph, a start node, an end node, and optionally a map of the coverage of each node, compute all simple paths between these nodes.

Modified from: https://github.com/Ninjani/rosalind/blob/e22ecf2c9f0935d970b137684029957c0850d63f/t_ba11b/src/lib.rs