Struct gfatk::gfa::graph::GFAdigraph

pub struct GFAdigraph(pub Graph<usize, (Orientation, Orientation, Option<i64>)>);

A wrapper of petgraph’s directed Graph struct, applied to a GFA. The edge weights included are the Orientation’s of the adjacent segments, and the coverage of this edge.

Tuple Fields
0: Graph<usize, (Orientation, Orientation, Option<i64>)>
Implementations

impl GFAdigraph

pub fn dot(&self, gfa: GFAtk) -> Result<()>

The main function called from gfatk dot.

It is a somewhat modified, simplified version of this: https://docs.rs/petgraph/latest/src/petgraph/dot.rs.html#1-349

Generating a DOT language output of a GFA file.

pub fn weakly_connected_components(
&self,
graph_indices: GFAGraphLookups
) -> Result<Vec<Vec<usize>>>

Split the GFA digraph into subgraphs which are the weakly connected components of the graph.

Taken from https://github.com/Qiskit/retworkx/blob/79900cf8da0c0665ac5ce1ccb0f57373434b14b8/src/connectivity/mod.rs

pub fn all_paths_all_node_pairs(
&self,
graph_indices: &GFAGraphLookups,
rel_coverage_map: Option<&HashMap<NodeIndex, usize>>
) -> Result<(Vec<(NodeIndex, Orientation)>, Vec<usize>, String)>

The main function called from gfatk linear.

This function will generate the longest path through the GFA, by filtering the output of all_paths, and choosing the path with the highest cumulative edge coverage.

pub fn node_count(&self) -> usize

Simple wrapper of Graph.node_count() in petgraph.

pub fn edge_count(&self) -> usize

Simple wrapper of Graph.edge_count() in petgraph.

pub fn trim(&self, graph_indices: GFAGraphLookups) -> Vec<usize>

Trim a graph to include only nodes connected to two or more other nodes.

This algorithm will loop for as long as the longest branch in the GFA yields a segment connected to only a single node.